Remove unnecessary byte casts in TagType

+++ src/main/java/org/glavo/nbt/tag/TagType.java
@@ -22,40 +22,40 @@ import java.util.function.Supplier;
 
 public final class TagType<T extends Tag> {
     /// 1 byte signed integer type. Sometimes used for booleans.
-    public static final TagType<ByteTag> BYTE = new TagType<>("TAG_Byte", (byte) 0x01, ByteTag.class, ByteTag::new);
+    public static final TagType<ByteTag> BYTE = new TagType<>("TAG_Byte", 0x01, ByteTag.class, ByteTag::new);
 
     /// 2 byte signed integer type.
-    public static final TagType<ShortTag> SHORT = new TagType<>("TAG_Short", (byte) 0x02, ShortTag.class, ShortTag::new);
+    public static final TagType<ShortTag> SHORT = new TagType<>("TAG_Short", 0x02, ShortTag.class, ShortTag::new);
 
     /// 4 byte signed integer type.
-    public static final TagType<IntTag> INT = new TagType<>("TAG_Int", (byte) 0x03, IntTag.class, IntTag::new);
+    public static final TagType<IntTag> INT = new TagType<>("TAG_Int", 0x03, IntTag.class, IntTag::new);
 
     /// 8 byte signed integer type.
-    public static final TagType<LongTag> LONG = new TagType<>("TAG_Long", (byte) 0x04, LongTag.class, LongTag::new);
+    public static final TagType<LongTag> LONG = new TagType<>("TAG_Long", 0x04, LongTag.class, LongTag::new);
 
     /// 4 byte floating point type.
-    public static final TagType<FloatTag> FLOAT = new TagType<>("TAG_Float", (byte) 0x05, FloatTag.class, FloatTag::new);
+    public static final TagType<FloatTag> FLOAT = new TagType<>("TAG_Float", 0x05, FloatTag.class, FloatTag::new);
 
     /// 8 byte floating point type.
-    public static final TagType<DoubleTag> DOUBLE = new TagType<>("TAG_Double", (byte) 0x06, DoubleTag.class, DoubleTag::new);
+    public static final TagType<DoubleTag> DOUBLE = new TagType<>("TAG_Double", 0x06, DoubleTag.class, DoubleTag::new);
 
     /// An array of bytes.
-    public static final TagType<ByteArrayTag> BYTE_ARRAY = new TagType<>("TAG_Byte_Array", (byte) 0x07, ByteArrayTag.class, ByteArrayTag::new);
+    public static final TagType<ByteArrayTag> BYTE_ARRAY = new TagType<>("TAG_Byte_Array", 0x07, ByteArrayTag.class, ByteArrayTag::new);
 
     /// A UTF-8 encoded string. It has a size, rather than being null terminated.
-    public static final TagType<StringTag> STRING = new TagType<>("TAG_String", (byte) 0x08, StringTag.class, StringTag::new);
+    public static final TagType<StringTag> STRING = new TagType<>("TAG_String", 0x08, StringTag.class, StringTag::new);
 
     /// A list of tag payloads, without tag IDs or names, apart from the one before the length.
-    public static final TagType<ListTag<?>> LIST = new TagType<>("TAG_List", (byte) 0x09, ListTag.class, () -> new ListTag<>(null));
+    public static final TagType<ListTag<?>> LIST = new TagType<>("TAG_List", 0x09, ListTag.class, () -> new ListTag<>(null));
 
     /// A list of fully formed tags, including their IDs, names, and payloads. No two tags may have the same name.
-    public static final TagType<CompoundTag> COMPOUND = new TagType<>("TAG_Compound", (byte) 0x0A, CompoundTag.class, CompoundTag::new);
+    public static final TagType<CompoundTag> COMPOUND = new TagType<>("TAG_Compound", 0x0A, CompoundTag.class, CompoundTag::new);
 
     /// An array of 4 byte signed integers.
-    public static final TagType<IntArrayTag> INT_ARRAY = new TagType<>("TAG_Int_Array", (byte) 0x0B, IntArrayTag.class, IntArrayTag::new);
+    public static final TagType<IntArrayTag> INT_ARRAY = new TagType<>("TAG_Int_Array", 0x0B, IntArrayTag.class, IntArrayTag::new);
 
     /// An array of 8 byte signed integers.
-    public static final TagType<LongArrayTag> LONG_ARRAY = new TagType<>("TAG_Long_Array", (byte) 0x0C, LongArrayTag.class, LongArrayTag::new);
+    public static final TagType<LongArrayTag> LONG_ARRAY = new TagType<>("TAG_Long_Array", 0x0C, LongArrayTag.class, LongArrayTag::new);
 
     /// Returns the tag type by its id; returns `null` if the id is invalid.
     public static @Nullable TagType<?> getById(byte id) {
@@ -87,9 +87,9 @@ public final class TagType<T extends Tag> {
     private final Supplier<T> tagSupplier;
 
     @SuppressWarnings("unchecked")
-    private TagType(String name, byte id, Class<?> tagClass, Supplier<T> tagSupplier) {
+    private TagType(String name, int id, Class<?> tagClass, Supplier<T> tagSupplier) {
         this.name = name;
-        this.id = id;
+        this.id = (byte) id;
         this.tagClass = (Class<T>) tagClass;
         this.tagSupplier = tagSupplier;
     }
